2.カスタムフック化
