+++ src/components/templates/todoTemplate/index.jsx
@@ -1,67 +1,24 @@
-import { useState, useMemo } from "react";
-import { defaultTodo } from "../../../common/index.js";
 import styles from './style.module.css'
 import { TodoList } from "../../organisms/TodoList"
 import { AddTodo } from "../../organisms/AddTodo"
 import { SearchArea } from "../../atoms/SearchArea"
+import { useTodo } from "../../../hooks/useTodo.js"
 
 export const TodoTemplate = () => {
 
-  // 表示用Todoリスト
-  const [showTodoList, setTodoList] = useState(defaultTodo);
-
-  // 検索用キーワード
-  const [searchKeyWord, setSearchKeyWord] = useState('');
-
-  // 追加用タスク
-  const [addInputValue, setAddTodo] = useState("");
-
-  // 採番Todoリスト
-  const [todoId, setTodoId] = useState(showTodoList.length);
-
-  // Todoリスト表示ロジック
-  const DisplayTodo = useMemo(() => {
-    return showTodoList.filter((todo) => {
-      const regexp = new RegExp("^" + searchKeyWord, "i");
-      return todo.name.match(regexp);
-    })
-  }, [showTodoList, searchKeyWord]);
-
-  // 検索用キーワードの変更
-  const onChangeSetSearchKeyWord = (e) => setSearchKeyWord(e.target.value);
-
-  // 追加用todoの変更
-  const onChangeAddTodo = (e) => setAddTodo(e.target.value);
-
-  // 追加todoロジック
-  const handleAddTodo = (e) => {
-    if (e.key == "Enter" && addInputValue != '') {
-      const nextTodoid = todoId + 1
-
-      const newTodoList = [
-        ...showTodoList,
-        {
-          id: nextTodoid,
-          name: addInputValue
-        }
-      ];
-
-      console.log("newTodoListの中身", newTodoList)
-      setTodoList(newTodoList);
-      // 採番IDを更新
-      setTodoId(nextTodoid);
-      // todo追加後、値をリセット
-      setAddTodo("");
+  const [
+    {
+      searchKeyWord,
+      addInputValue,
+      DisplayTodo
+    },
+    {
+      onChangeSetSearchKeyWord,
+      onChangeAddTodo,
+      handleAddTodo,
+      deleteTodo
     }
-  }
-
-  // todo削除
-  const deleteTodo = (targetId, targetName) => {
-    if (window.confirm(`${targetName}のタスクを削除しますか`)) {
-      const afterDeletedTodoList = showTodoList.filter(list => list.id !== targetId);
-      setTodoList(afterDeletedTodoList)
-    }
-  }
+   ] = useTodo();
 
   return (
     <div className={styles.container}>
